AsyncInflateConfig.java
